amend requested changes
