+++ src/atoms/ContextMenu.ts
@@ -1,4 +1,5 @@
-import { atomWithReset } from "jotai/utils";
+import { useAtom } from "jotai";
+import { atomWithReset, useResetAtom } from "jotai/utils";
 
 export type SelectedCell = {
   rowIndex: number;
@@ -14,5 +15,16 @@ const INIT_VALUE = {
   selectedCell: null,
   anchorEl: null,
 };
+
+export default function useContextMenuAtom() {
+  const [contextMenu, setContextMenu] = useAtom(contextMenuAtom);
+  const resetContextMenu = useResetAtom(contextMenuAtom);
+
+  return {
+    contextMenu,
+    setContextMenu,
+    resetContextMenu,
+  };
+}
 
 export const contextMenuAtom = atomWithReset<IContextMenuAtom>(INIT_VALUE);
